requirements.txt
git+https://github.com/benob/mosestokenizer.git
numpy
regex
torch
tqdm
transformers

pyttsx3
soundfile
PyAudio
sounddevice
elevenlabslib

requests
openai-whisper
vosk

obsws-python

azure-cognitiveservices-speech